refactor(sdk): host inferProvider and supported-provider list; langchain re-exports
