+++ packages/langchain/src/model-provider-resolver.ts
@@ -1,35 +1,9 @@
 import type { BuiltInModelProviderId, ModelProviderId } from "@dawn-ai/sdk"
+import { inferProvider, SUPPORTED_AGENT_PROVIDERS } from "@dawn-ai/sdk"
 
-export const SUPPORTED_AGENT_PROVIDERS = [
-  "openai",
-  "anthropic",
-  "google",
-  "mistral",
-  "groq",
-  "ollama",
-  "xai",
-  "openrouter",
-] as const satisfies readonly BuiltInModelProviderId[]
+export { inferProvider, SUPPORTED_AGENT_PROVIDERS }
 
 const supportedProviderSet = new Set<string>(SUPPORTED_AGENT_PROVIDERS)
-
-export function inferProvider(model: string): BuiltInModelProviderId | undefined {
-  const normalized = model.trim().toLowerCase()
-
-  if (/^(gpt-|o3|o4)/.test(normalized)) return "openai"
-  if (normalized.startsWith("claude-")) return "anthropic"
-  if (normalized.startsWith("gemini-")) return "google"
-  if (
-    normalized.startsWith("mistral-") ||
-    normalized.startsWith("mixtral-") ||
-    normalized.startsWith("codestral-")
-  ) {
-    return "mistral"
-  }
-  if (normalized.startsWith("grok-")) return "xai"
-
-  return undefined
-}
 
 export function resolveProvider(options: {
   readonly model: string
