+++ packages/sdk/src/index.ts
@@ -32,6 +32,7 @@ export type {
   BuiltInModelProviderId,
   ModelProviderId,
 } from "./model-provider.js"
+export { inferProvider, SUPPORTED_AGENT_PROVIDERS } from "./model-provider.js"
 export type { RouteConfig, RouteKind } from "./route-config.js"
 export type { RouteStateMap, RouteToolMap } from "./route-types.js"
 export type {
